refactor(backend): isolate Skill marketplace domain
