+++ tests/architecture/rustModuleBoundaries.test.ts
@@ -86,12 +86,13 @@ describe("Rust modular architecture boundaries", () => {
 
     expect(provider).toContain("mod universal;");
     expect(skill).toContain("mod discovery;");
+    expect(skill).toContain("mod marketplace;");
     expect(proxy).toContain("mod takeover;");
     expect(codex).toContain("mod storage;");
 
     for (const source of [provider, skill, proxy, codex]) {
       expect(source).not.toMatch(
-        /pub(?:\(crate\))? mod (?:universal|discovery|takeover|storage);/u,
+        /pub(?:\(crate\))? mod (?:universal|discovery|marketplace|takeover|storage);/u,
       );
     }
   });
